requirements.txt
certifi==2020.12.5
chardet==3.0.4
geographiclib==1.50
geopy==2.0.0
googlemaps==4.4.2
idna==2.10
openrouteservice==2.3.0
paho-mqtt==1.5.1
pyzmq==20.0.0
requests==2.25.0
rstr==2.2.6
urllib3==1.26.2
